extract util func
