+++ packages/ui/package.json
@@ -10,7 +10,8 @@
     "@radix-ui/react-popover": "^1.0.6",
     "@radix-ui/react-select": "^1.2.2",
     "@radix-ui/react-tabs": "^1.0.4",
-    "clsx": "^2.0.0"
+    "clsx": "^2.0.0",
+    "@ossinsight/widgets-utils": "workspace:^"
   },
   "devDependencies": {
     "@faker-js/faker": "^8.0.2",
