+++ packages/ui/src/components/Selector/TimeZoneSelector.tsx
@@ -1,23 +1,9 @@
 import * as React from 'react';
 
 import { useSimpleSelect } from './Select';
+import { generateZoneOptions } from '@ossinsight/widgets-utils/src/ui';
 
-const ZONE_OPTIONS: Array<{ key: number; title: string }> = [];
-
-for (let i = -12; i <= 13; i++) {
-  ZONE_OPTIONS.push({
-    key: i,
-    title: i > 0 ? `+${i}` : i === 0 ? '0' : `${i}`,
-  });
-}
-
-const DEFAULT_ZONE = Math.max(
-  Math.min(
-    Math.round(12 - new Date().getTimezoneOffset() / 60),
-    ZONE_OPTIONS.length - 1
-  ),
-  0
-);
+const { DEFAULT_ZONE, ZONE_OPTIONS } = generateZoneOptions();
 
 export interface TimeZoneSelectorProps {
   onValueChange?: (newValue: string) => void;
